starting date 2008 q4, before which GTA data is reporting only 0's, due to lack of entries at that time
now code check ready

important: slight differences of +2 +3 here and there, nothing shocking, but this goes back as far as even 2009. Larger differences (in comparison to what was sent to Mr. Hjersbech last time) in 2019 Q3 which would be expected, the numbers now seem more coherent with the evolution over time